// UiPreparation/UI/src/app/models/user.ts
export class User {
    userId: number = 0;
    name: string = "";
    surname: string = "";
    account: string = "";
    email: string = "";
    mobilePhones: string = "";
    address: string = "";
    notes: string = "";
    status: boolean = false;
    gender: number = 0;
    refreshToken: string = "";
    verified: boolean = false;
}